Move NA values and related functions to a new file
