+++ rdata/parser/__init__.py
@@ -2,8 +2,6 @@
 
 from ._parser import (
     DEFAULT_ALTREP_MAP as DEFAULT_ALTREP_MAP,
-    R_FLOAT_NA as R_FLOAT_NA,
-    R_INT_NA as R_INT_NA,
     CharFlags as CharFlags,
     RData as RData,
     RExtraInfo as RExtraInfo,
@@ -11,7 +9,6 @@ from ._parser import (
     RObjectInfo as RObjectInfo,
     RObjectType as RObjectType,
     RVersions as RVersions,
-    is_float_na as is_float_na,
     parse_data as parse_data,
     parse_file as parse_file,
 )
